requirements.txt
requests
transformers
python-dotenv
flask